Various cast fixes and guards

+++ innd/rc.c
@@ -574,7 +574,7 @@ RCreader(CHANNEL *cp)
         /* Use cp->Rejected for the response code in CHANclose. */
         new->Rejected = reject_val;
         RCHANremove(new);
-        WCHANset(new, reject_message, (int) strlen(reject_message));
+        WCHANset(new, reject_message, strlen(reject_message));
         WCHANappend(new, RCterm, strlen(RCterm));
         WCHANadd(new);
         free(reject_message);
@@ -648,7 +648,7 @@ RCreader(CHANNEL *cp)
         /* Use cp->Rejected for the response code in CHANclose. */
         new->Rejected = reject_val;
         RCHANremove(new);
-        WCHANset(new, reject_message, (int) strlen(reject_message));
+        WCHANset(new, reject_message, strlen(reject_message));
         WCHANappend(new, RCterm, strlen(RCterm));
         WCHANadd(new);
         free(reject_message);
